feat(akari/ent): Create a database schema and migration (#35)
* chore(akari/internal/ent): Create a empty system prompt schema

Generate a code with ent to implement.

* feat(akari/internal/ent): Fill the system prompt schema

Add required fields to be used with AI.

* feat(akari): Create database migration files with Atlas

Generate and migarte with Makefile.

* fix(akari): Update the Atlas config PostgreSQL version to 17

It's still 15, the previous version.

* style(akari/ent): Move the ent schema package

Move to the root of the repository.

* fix(akari): Correct the ent. schema path

It was moved but not reflected.

* chore(akari/ent): Remove the version field

No need to have it with these fields.

* chore(akari/ent): Remove the previous_prompts field

It's not used since there's no version control.

* chore(akari): Add Atlas hash regeneration command to Makefile

Keep the hash up-to-date.